apps/ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pymongo==4.6.0
motor==3.3.2
redis==5.0.1
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
prophet==1.1.5
xgboost==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
joblib==1.3.2

# AWS Dependencies
boto3==1.34.0
botocore==1.34.0
awscli==1.32.0

# Additional ML libraries
scipy==1.11.4
